// frontend/src/features/not-found/types/not-found.types.ts
import type { LucideIcon } from "lucide-react";

export type NotFoundFeatureCard = {
  icon: LucideIcon;
  iconColorClass: string;
  darkIconColorClass: string;
  bgColorClass: string;
  darkBgColorClass: string;
  gradientClass: string;
  hoverBgClass: string;
  darkHoverBgClass: string;
  title: string;
  description: string;
  to: string;
};